Made tweaks to lighting
Changed render mode of Camera Light to "Important" so it wouldnt affect the other lights in the scene as much, but its still a little wonky, we need to find a way to made the lights not interact with each other.

Added an Editor Light which can be toggled while making edits in scene view so u can see everything while editing.