Setup random level generation
Definitely not perfect and will need some tweaking, the only real stipulation rn is that enemies and the exit aren't allowed to spawn within a certain radius of the player and I want to set it up a bit more specific so like there's no chance of things overlapping and the space gets filled better, but it works.
Has functionality to specify how much bigger the map should be between levels, and how many more enemies/obstacles/powerups/etc there should be in each subsequent level.
It should be easy to drag in and drop powerups and stuff as we make them and they should spawn in randomly.